[SCRUM-95] refactor: 공통 응답 디렉토리 수정
