+++ src/main/java/com/projectlyrics/server/domain/artist/service/ArtistQueryService.java
@@ -1,7 +1,7 @@
 package com.projectlyrics.server.domain.artist.service;
 
 import com.projectlyrics.server.domain.artist.dto.response.GetArtistResponse;
-import com.projectlyrics.server.domain.common.dto.CursorBasePaginatedResponse;
+import com.projectlyrics.server.domain.common.dto.util.CursorBasePaginatedResponse;
 import org.springframework.data.domain.Pageable;
 
 public interface ArtistQueryService {
